bugget 改为float 评价回复ui修改

+++ LetsgoGuideApp/app/src/main/java/com/njz/letsgoappguides/model/home/OrderDesignInfo.java
@@ -242,7 +242,7 @@ public class OrderDesignInfo implements Parcelable {
         return bugGet;
     }
 
-    public void setBugGet(int bugGet) {
+    public void setBugGet(float bugGet) {
         this.bugGet = bugGet;
     }
